Compiler warnings cleanup
Refactored all code so that compiles under all flavors of gcc compiler
issue no (or almost no) warnings.

Fixed undexta to handle input file names with extension correctly.

Fixed a benign buffer overflow in QV.c

+++ dexta.c
@@ -191,7 +191,7 @@ int main(int argc, char *argv[])
                     break;
                   rlen += x;
                   if (rlen + MAX_BUFFER > rmax)
-                    { rmax = 1.2 * rmax + 1000 + MAX_BUFFER;
+                    { rmax = ((int) (1.2 * rmax)) + 1000 + MAX_BUFFER;
                       read = (char *) Realloc(read,rmax+1,"Reallocaing read buffer");
                       if (read == NULL)
                         exit (1);
